Remove conversion warnings.

+++ src/parser.c
@@ -127,7 +127,7 @@ PARSER_status_t PARSER_get_parameter(PARSER_context_t* parser_ctx, STRING_format
         goto errors;
     }
     // Convert string.
-    string_status = STRING_string_to_integer(&((parser_ctx->buffer)[parser_ctx->start_index]), format, param_size_char, parameter);
+    string_status = STRING_string_to_integer(&((parser_ctx->buffer)[parser_ctx->start_index]), format, (uint8_t) param_size_char, parameter);
     STRING_exit_error(PARSER_ERROR_BASE_STRING);
     // Update start index after decoding parameter.
     if ((parser_ctx->separator_index) > 0) {
